refactor(costs): extract plan detection into src/lib/plan.ts (#432)
Closes #161 (partial - first extraction step)

Moves PlanType, PlanDetection, detectPlan, getPlanType, isMaxPlan,
getPlanDescription out of costs.ts (1202 lines) into a focused
src/lib/plan.ts module. costs.ts re-exports for backwards compat.

Reduces costs.ts from 1202 to 1118 lines (-84 lines).
All existing imports from costs.ts continue to work unchanged.

Also fixes git.test.ts GIT_DIR hook recursion bug by clearing
GIT_DIR and GIT_WORK_TREE in beforeEach.


Agent: cli/issue-solver
Squad: cli
